chicken_health_api/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
torch>=2.2.0
torchvision>=0.15.2
timm>=0.9.5
Pillow>=10.0.1
numpy>=1.24.3
pydantic>=2.4.2